+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Change Log
 
 
+## Unreleased
+
+### Added
+
+- **archive/tar:** A Reader which reads a certain file from a TAR archive, optionally decompressing it
+- **arcive/tar:** `NewTarGzFileReader` (returns a Reader which decompresses and unarchives a .tar.gz stream and returns a file from it)
+
+
 ## 0.0.2 - 2017-04-03
 
 ### Changed
@@ -12,4 +20,4 @@
 
 ### Added
 
-- **str:** `ToSnake` util to convert string to snake case
+- **strings:** `ToSnake` util to convert string to snake case